BUG: Rename remote module FDFImageIO -> IOFDF
The module's name is inconsistent and causing it not to be usable as a
module in ITK. Previously ITK was calling it FDFImageIO, while
internally it was called ITKFDFImageIO. Neither name followed ITK
conventions. The correct name for an external module does not begin
with the ITK prefix and for this case should be: IOFDF

The name of the module and related variable, libraries, export
specifications etc, have been rename consistently.

Change-Id: Idf4267aaf1c89859924d301ea137dea9b0c269cc