COMP: Add c++11 override designation
C++11 provides the override keyword to assit with identifing
functions that are meant to be override the base class
implementation.  This refactoring adds the override keyword
in order to enhance compile time identification of
errors that may be introduced during refactoring.

The following substitutions were heavily used during
the code instrumentation of the override keyword:

:s/)$/) ITK_OVERRIDE/ge
:s/);$/) ITK_OVERRIDE;/ge
:s/) *const *;$/) const ITK_OVERRIDE;/ge
:s/const *{/const ITK_OVERRIDE {/ge
:s/) *const$/) const ITK_OVERRIDE/ge
:s/() *{}$/() ITK_OVERRIDE {}/ge
:s/) = 0;$/) ITK_OVERRIDE = 0;/ge
:s/) *const *= *0;$/) const ITK_OVERRIDE = 0;/ge
:s#();  *//#() ITK_OVERRIDE;  //#ge

Change-Id: I199c677b6420831ada5b009fc3d222b00687321a