BUG: Enable setting all zeroes for fixed parameters for displacement
This initial state for the displacement field is with all zeros for
fixed parameters and null for the displacement image. If a clone was
done an a default constructed displacement field transform and
exception was generated due to 0-spaceing.

This patch adds a special case to set the initial null state with
zeros for fixed parameters.

Change-Id: Id6a75d112313356db184a37eb1e2cfab5072129d