BUG: Return a correct voronoi partition in Danielsson filter example.
Use the ConnectedComponentImageFilter to label each group of pixels.

Adds a new image of FivePoints.png which is the original with each pixel dilated
out, which is more useful in the ITK Software Guide. Keep the original for the
tests.

There may be a regression in the DanielssonDistanceMapImageFilter. The
SetInputIsBinary() method does not work as expected, causing the
GetVoronoiMap() function to return a blank image. The voronoi map
is returned blank if the VoronoiPixelType is not unsigned char, so I've
changed the example to reflect that.

Change-Id: I6bdc40b0314c3cc4a3261ee2c1f2d1f1e00cf794