BUG: Clear TransformIO's list of transforms in TransformFileReader.
Since TransformFileReaderTemplate and TransformIO do not have very good const
correctness and encapsulation, it is possible to manipulate their members and
change their state in a way they might not expect.

In 81faafe29, a m_TransformIO member was added.  This is used in
TransformFileReaderTemplate's Update() step instead of creating a new
TransformIOBase instance on every Update.  To keep completely consistent
with the previous behavior, we must clear the TransformIOBase's
internal transform list at the time the class was previously instantiated.

Thanks to Jean-Christophe Fillion-Robin and the Slicer community for
identifying this change in behavior.

A note is also added indicating that the output is not intended to be
modifiable and the non-const version was deprecated.

Change-Id: If5dfd03b32b31ba765c6781d689e91edb944be3f