ENH: Check regions in HDF5 streaming test
This commit make use of PipelineMonitorImageFilter in
itkHDF5ImageIOStreamingReadWriteTest.cxx to check that the regions written and
read to/from file with streaming are as expected.

It shows that the streaming correctly works when reading HDF5 file, but do
no work when writing file, because the writer input is not streamable.

Change-Id: Icda6c4d6b6afbfb278f7dcaef5dedf90999b7713