ENH:  Making the point set type visible.
The current code set for the image registration methods
hard code the fixed and point set types to have a pixel
type of unsigned int.  This is useful for labeled point
sets for use with the LabeledPointSetToPointSetMetric.
However, this should be generalized to allow for other
point set metric types including non-labeled point sets.
Exposing the point set type is the first step to this
generalization.

Change-Id: Ifc0441dccdce7794c34b912e32db3218cd3a7e0d