ENH: Modularize KWStyleTests
The patch extends the KWStyleTest to the itk_module macro. Thereby,
adding KWStyle checks to external and remote modules. Additional
arguments could be added to the itk_macro or per module configuration
file to change the KWStyle options per module.

Each module can now have a ITKKWStyleOverwrite.txt and
ITKKWStyleFiles.txt.in, to specify the style exceptions and the files to
run KWStyle on.

The make target StyleCheckCode has been removed.

Change-Id: Ia5c7e3dd33ae18bb7220f92a6641c4f3ce410031