ENH: Improve wrapping for Vector image types
Defines the covariant vector image types in WrapITKTypes instead of doing
it in itkImage.wrap

Add more vectorial types to the wrappings. The vectorial dimensions
are the same as the image dimension (this is the same behaviour as before).
A separate list for vectorial dimension would need to be
added in a further patch.

Change-Id: I1426e30138b451e1cee7f53d7ebc9e02bf0684f5