ENH: Add Get/SetSigmaArray to GradientRecursiveGaussianImageFilter
When wrapping the GradientRecursiveGaussianImageFilter with vectorial
input, swig did not like the fact that GetSigma would return either
a double, or an array.

This patch adds a getter and a setter for the sigma array.
It is inspired by the code in:
Filtering/Smoothing/itkSmoothingRecursiveGaussianImageFilter,
where the same strategy was used.

Note that like in the itkSmoothingRecursiveGaussianImageFilter,
the itkConceptMacro needed to be disabled for it to work with
wrapping.

Change-Id: I74a7961136d0fb9c8a2a9b7eafafb10c67114d5d