ENH:  Cleaning up point set reg. in BSplineSyN
The PointSetToPointSetMetricv4 class is set up to handle both
global and local transforms.  For the latter, this involves
returning a dense metric derivative where the only non-zero
values are located at the fixed point locations (albeit with
a round-off error since the point sets have to be mapped to
the closest index).  Although this provides a generic interface
for dense transforms, it might not be ideal for all such
transforms.  For example, the B-spline SyN transform can
accommodate sparse derivatives at continuous locations within
the displacement field domain while not requiring the extra
processing of the filler zero derivatives.

Change-Id: I6488fbd4f35e44b25c111b3b83877c0cae5458d0