BUG: Suppress wrapping type conversions warnings with MSVC.
These are usually suppressed by the pragma's in itkWin32Header.h. In
wrapping, the *Python.cpp files are generated by swig and the content and
order of the content is determined by swig. In the generated content,

  #include <stdexcept>

comes be for the *SwigInterface.h, which include the ITK include's.  For
the pragma to take effect, it must become before the the first STL header
is include'd.

A flag to suppress the warning is added to the command line when building
the Python module.

This addresses:

Building [...] Wrapping\Modules\ITKSpatialObjects\...SpatialObjectsPython.dir\itkDTITubeSpatialObjectPointPython.cpp.obj

[...] C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\INCLUDE\xutility(2045) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data

Change-Id: Ie339fd5e2084c1b6c7344fea59a734b329cec15e