COMP: fix warning with non default template parameter for OutputImage
LabelType is hard coded to be SizeValueType, but m_OutputBackgroundValue is
OutputPixelType. Resulting in a warning when the expected LabelType is not
SizeValueType. Instead of constraining LabelType to be SizeValueType, we deduce
it from the template argument.

public typedef for LabelType has been removed, instead it has been replaced by
InternalLabelType (i.e. SizeValueType) and the actual output type for labels
are now deduced from template argument (i.e. OutputPixelType).

Change-Id: I70b813871e902052d20d9e5c09bc0c3239c5e6f2