BUG: Update AmoebaOptimizer NumberOfIterations
The number of iterations was not being updated on iteration
events. Additionally the number of iterations wasn't being reset to
zero when run multiple times.

The WithRestarts loop have been updated to use the continuously update
m_CurrentIteration IVAR to limit the number of restarts and use the
m_NumberOfIterations as the total limit on iterations/evaluations.

Change-Id: Ie37c34b98dd0f5794dc2bbf28f8c76fa2e901bf7