COMP: _Pragma suppress fltpnt equality warnings
The clang compiler has good testing for floating point equality
testing.  In some cases it is OK to test exact equality, but it is
often a an indication that in-equality with an epsilon would be a
more robust and stable solution.

The clang pragmas are added to allow for suppression of false
positives when exact equality is really desired.

Use _Pragma for defining pragmas inside macros.

Change-Id: Id154d451740a90acf1af146a03ea3c32a8a9fdf5