BUG: ImageType, CofficientImageArray types missing
Only the BSplineTransformType defines the extra
typedefs for ImageType and CofficentImageArray.
The BSplineTransformType needs to be explicitly
used for the BSplineTransform overloaded
function.

Building with explicit instantiation identified this
problem.  The overloaded functions are compiled
for every transform type, even though they can not be
accessed.  The BSplineTransformType has necessary
typedefs that are not available in the other transform
types.

Change-Id: I59547ff964ba77044205c392cfcf90306a9f118b