ENH: Add ProcessObject::AddOptionalInputName and more named inputs
Add ProcessObject::AddOptionalInputName method which replaced
duplicate and unclear approach of Adding a RequiredNamedInput followed
by removing, so that the named input entry will exists but will not be
marked as required.

This method was then used in many filter which have optional inputs,
and the appropriate name is also added to the inputs.

Change-Id: Ide4569ac4d54db218f104746c8e12bfa89b42db5