ENH: let the user choose the tradeoff between size and speed in fft operations
fft transformations usually have a constraint on the size of the input. That
constraint can be a hard constraint - the transform can't be run when it is
not valid - or a soft constraint - the transform runs but the performance may
be dramatically affected.

That constraint usually depends on the greatest prime factor of the
decomposition of the size of the input image. For vnl implementation, the hard
limit is 5, and 3 and 2 or valid values.
FFTW has no hard limit, but prime factor above 13 has a dramatic effect on the
performance.
Some other implementations may be limited to 2.

A similar mechanism was present in the original "FFT based convolution" paper.
A virtual method has be added to all the fft filters so the user can retreive
their constraint at runtime. Hopefully this should help also when we'll do the
fft filter selection with a factory.

Change-Id: I0ed44f1af9f69f1077c2cabb0af5d37253ae3781