BUG: Link to VTK 6 rendering backend modules.
Some projects that use both VTK and ITK present compile time warnings
like: warning: "vtkRenderingCore_AUTOINIT" redefined [enabled by
default].

Projects that link against an ITK that turns on modules with VTK
dependencies exhibit this bad behavior.

These warnings are caused by compile time defines
-DvtkRenderingCore_AUTOINIT="3(vtkInteractionStyle,vtkRenderingFreeType,vtkRenderingOpenGL)"
and
-DvtkRenderingCore_AUTOINIT="4(vtkInteractionStyle,vtkRenderingFreeType,vtkRenderingFreeTypeOpenGL,vtkRenderingOpenGL)"

VTK does not yet bring in required VTK rendering backend dependent modules.
We explicitly link to vtkRenderingFreeTypeOpenGL so compatible defines are
created and warnings and link errors are avoided.

Also, support for the OpenGL2 backend is added (available in 6.2.0 and later)
by using the VTK_RENDERING_BACKEND variable that is defined in
VTKConfig.cmake.

Missing checks for VTK_VERSION in the ITKLevelSetsv4Visualization module
export code are also added.

Only link the vtkRenderingFreeTypeOpenGL library if it is available.  It was
removed in VTK master (to be released as VTK 6.3).

Thanks to Jean-Christophe Fillion-Robin for initiating the discussions and
Bill Lorensen for the initial implementation and description.

Change-Id: I85547df5e58968c6c7dc8b3294c9308ab29bb28a