BUG: GDCM set default ultrasound spacing when not found.
From the bug report ITK-3360.

When the following conditions (at line 422 of itkGDCMImageIO.cxx):

    std::vector<double> sp;
    gdcm::Tag spacingTag(0x0028,0x0030);
    if(ds.FindDataElement(spacingTag) &&
     !ds.GetDataElement(spacingTag).IsEmpty())

are false sp will be unitialized and used at the following lines:

     spacing[0] = sp[0];
     spacing[1] = sp[1];
     spacing[2] = 1.0; // punt?

Only use the spacing when it is available, otherwise use 1.0. Add a new test
with an input file that does not have any spacing. Without this fix the new
test will segfault.

Change-Id: I65a2625b4b672efb56001f108075b099145c5c6f