ENH: Add itkFFTPadImageFilter
This code was adapted from the Insight Journal contribution:
"FFT Based Convolution" by Gaetan Lehmann
http://hdl.handle.net/10380/3154

It complements previously integrated classes.

The main change from the original contribution is the removal of the kernel
input in order to compute the required padding for a FFT based convolution.
While useful, this feature is complexifying the code and the API, and makes
the assumption that FFT is mostly useful for convolution.
It may be implemented later in another filter.

Change-Id: I5b22a03499998c4c2ef34a939f82ad7b81896575