COMP: Run the SWIG configure step in a CMake script.
The SWIG configure step yields a false-positive warning on CDash when
CTEST_USE_LAUNCHERS is on. Work around this by running the configure step in a
CMake script and forwarding the result value from running the configure script
with execute_process.

Based off the suggestion from Jean-Christophe Fillion-Robin and the scripts
used in Slicer.

Remove the old warning suppression that was not effective.

Change-Id: If4a908b55f5c7cc75ab4233e4495179dfdc8bc8a