BUG: Use consistent SizeGreatestPrimeFactor in deconvolution tests.
The deconvolution filters inherit from the FFTConvolutionImageFilter and their
behavior will exhibit numerical differences when a different
SizeGreatestPrimeFactor is used to compute the FFT's because of different
amounts of padding.  FFTW and VNL backends have different
SizeGreatestPrimeFactor.  Set the SizeGreatestPrimeFactor, which both backends
are capable of, so we get consistent test results.

Change-Id: Id863107b6096318722e2c37a8811c896c9f20ebc