COMP: Build failure using EuclideanDistancePointMetric with mismatched meshes.
itk::PointSetToPointSetMetric is templated over fixed and moving point set
types.  Since itk::Mesh and itk::QuadEdgeMesh inherit from itk::PointSet,
itk::PointSetToPointSetMetrics should be able to compare these types.
This works as expected when both moving and fixed meshes are of the same type.
However, the project fails to build when the types are mismatched (specifically
when one type is an itk::Mesh and the other is an itk::QuadEdgeMesh).  This
is due to the fact that itk::PointSetToPointSetMetric defines only a single
iterator for iterating over both fixed and moving types.  Adding separate
fixed and moving iterators fixes this bug.  A test has been provided which
fails to build in the absence of this fix.

Change-Id: I90b57d1730cd4abb8eb2bdd888b7323387227e20