ENH: Add AtmoicInt class, use TimeStamp & LightObject
Import AtomicInt class implementation from VTK. Changing style to ITK,
enabling 64-bit atomic integers on OSX and linux on 32-bit
architectures.

Use AtomicInt in LightObject for reference counting. The integer type
was only exposed as an "int" type in the public interface. The
internal type is now just an int. The internal methods used for the
Windows and OSX atomic operations remain the same but for GCC the are
changed from __atomic_add to __sync_add_and_fecth .

Use AtomicInt in TimeStamp.

Change-Id: I82d6e2ed16020a3d97b17d7e7c7a260bbcae7da1