ENH: Add post-commit hook for third party library patches.
This commit adds a post-commit hook to look through the contents of a patch.
If files in third party libraries are modified, instructions are given on how
to create a patch that can be submitted upstream.

For third party modules that use the Git subtree UpdateFromUpstream.sh
script, URLs with contribution instructions are also presented to the user.
Additional URLs should be added here as more third party modules are converted
to the subtree update script mechanism.

The SetupForDevelopment version is bumped to force an update to the hooks
repository in .git/hooks, which has the post-commit hook chain.

Change-Id: I563b65d2d34e18a7a5882bd7f0f4cb4d5f2dd47b