ENH: Use gold linker on Linux with GCC.
The gold linker is approximately three times as fast as the default linker.
Enable it with GCC 4.8 and newer where it was tested. The gold linker only
works with ELF executables, so it is only enabled for Linux.

Timing tests:

Static libs link time:

  ld:      452 sec
  ld-gold: 179 sec

Shared libs link time, wrapping enabled:

  ld:      164 sec
  ld-gold: 48 sec

Change-Id: Ic8d9b5132ad6b71e5f100b5436615d6a8fac3e5c