ENH: Add ITKModuleExternal.cmake
This allows a module to be built outside the ITK source tree by adding the
following to the module's top level CMakeLists.txt:

  project(MyModule)

  if(NOT ITK_SOURCE_DIR)
    find_package(ITK REQUIRED)
    list(APPEND CMAKE_MODULE_PATH ${ITK_CMAKE_DIR})
    include(ITKModuleExternal)
  else()
    itk_module_impl()
  endif()

Move CreateTestDriver.cmake and itk_add_test into CMake/ITKModuleTest.cmake so
they can be re-used.

Fix bug in the ModuleCPPCheckTest.cmake module -- the file(GLOB ...) command
was not checking the correct directory. Also use ITK style for the test name.

Support for the include, src, and test directories was added.  Support for
wrapping is only partially implemented and will be completed in a later
commit.

This cannot be used for building against an ITK build tree, not an install
tree.

Change-Id: If83d5fd265ca0985a3fff4a511a3a3e3a659bae7