BUG: Remove shared specifications from TransformIO
The TransformIO is not yet instrumented to properly
deal with shared libraries.  This will need investigation
and modifications similar to what was needed for
ImageIO libraries.

A close inspection of the itk::ImageIO mechanisms for
dealing with shared libraries needs to be completed,
but that effort will need to wait until after the next
release is performed.

Change-Id: I14d7bc596c1252f1461772f8e761214b628dba8e