BUG: Remove reliance on exceptions in StringTools::ToData methods
GCC 5.1 appears to have a bug with regards to the
std::ios_base::failure symbol. Using the headers results in a cxx11
abi specified symbol, why those emitied from libstdc++ are missing the
abi specification. This causes catch statement not to match types, as
the run-time type information does not match.

The offending code practice was re-written to check the state of the
stream and utilize ios_base::exceptions to throw the problematic
exception after it's determined it is not a normal eof state.

Remove clear, and seek before destruction of stream objects. This should
have no effects.

Change-Id: Ib1e6405bbb19d873d7cfbc2d02d9fc414d9e8055