DOC: Document grafting the input image in composite filters.
This is a best practice that prevents redundant upstream pipeline updates.

The first filter in a mini-pipeline is connected to a new image pointer that has
grafted pipeline input.

Document in the example that is used by the Software Guide and in the
ImageSource doxygen documentation.

Also make style improvements to the CompositeFilterExample.

Suggested-by: Bradley Lowekamp <blowekamp@mail.nih.gov>

Change-Id: If15564e7aabeff0e33218db2da7379cf35fea16d