PERF: Performance enhancements on VariableLengthVectors 1/...
VariableLengthVector pixels induce many inefficiencies on algorithms and
filters working on VectorImages.

This commit is the first of a series on the subject.
More explanations on what's to come are available on the draft page
(http://wiki.orfeo-toolbox.org/index.php/OptimizingVectorImageProcessing).

tl;dr: the cost of transformations like LinearInterpolateImageFunction on
monoband itk::VectorImage<> shall be as close as the cost of the same
processing while the image is loaded as a itk::Image. And it can be done.

First things first.
Today implementation of VLV assignment operator induce two things that 99% of
the time never make sense:
- we only need to allocate memory the first time -- within loops this cost
  makes no sense; I assume the size of a pixel is stable.
- we never need to keep old values

Moreover, there wasn't any version of SetSize() that permits to avoid
reallocations or to duplicate values when we know we don't need that.

The choices I've made here are:
- To provide a very-low cost overload of SetSize() that permits end user to
  choose without any ambiguity what he exactly wants to acheive.
  Hence the two policies about Reallocations, and about KeepingOldValues.
- The behaviour of the current overload hasn't been changed
- The code has been factorized, and IMO simplified (this may be subjective).

The difficulties I had are related to the fact that sometimes VLV are acting as
proxies (to memory hold elsewhere). I've tried to express the contracts (as in
Design By Contract) of the various operations that implies reallocations or
assignments.

In an ideal world, we should not be permitted to assign values to Proxies. In
other words, proxies should be restricted to types like
VariableLengthVector<const double> instead of VariableLengthVector<double>.
This way, we could have easily asserted that the size of proxy VLVs cannot be
changed, and that proxy VLV could not be assigned. I'm afraid it's to late to
proceed to such radical changes.

Notes:
* I've proceeded to other optimizations. For instance, I've got rid of the
auto-assignment pessimistic test. As the assignment is written to provide a
Strong Exception Guarantee, the only thing the auto assignment test does is to
have a us pay for a test that tries to optimize a situation that seldom, if
never, occur in normal code. Thanks to the Strong Exception Guarantee (assuming
that copy operations on the value type are noexcept), the code will resist to
auto-assignment.

* I've hesitated to provide move assignment and move copy. It would require
some testing and macros, but it should be easy to add. For another patch may
be?

Change-Id: I0b07c5157cf5b8ae1b5b1de1181d7db4364596ed