PERF: Reworked MMI Threading using a per-thread buffer
Supporting data for this topic can be found on Google Docs.

Mattes Mutual Information Performace Improvements:
https://docs.google.com/document/d/1W9DCGZojbhsMy8E3KA8wFu7UObAx2ha9cEGAgcKp4OI/edit?usp=sharing

MMI Test Data:
https://docs.google.com/spreadsheets/d/1ZGcSg7Dh4J_o2K0CMOG8Fqnwidz8EthhjaN5mu_8KEo/edit?usp=sharing

The current threading implementation of MMI uses a vector of
three-dimensional Image objects to be created when trying to get the
derivative. The number of threads currently being used determines the
size of this vector and these objects are expensive to create. The
bigger performance drain though is the need to then go through each
of these Image objects and combine the values from each object into
one accumulated Image object. This method wastes both time and memory.
Since the vast majority of the values in these Image objects are never
actually used and remain at zero the time spent creating them and then
going through them value by value is largely wasted. On top of this
the memory space used is also being wasted.

The new implementation replaces this vector of Image objects with a
vector of buffers. The number of threads also determines this vector’s
size. Each buffer will store the value and offset of each operation
that needs to be performed on the final accumulated Image object. This
means that only operations that are actually doing work are executed
and memory space isn’t wasted holding zero values. This also gives the
opportunity to periodically attempt to apply the operations stored in
the buffer to the accumulated Image object during threading further
improving performance.

The new buffer manager can be found in
itkMattesMutualInformationImageToImageMetricv4.h. Testing shows that
this new buffer improves performance for higher thread counts, usually
four or more threads, and makes memory usage fairly constant rather
than growing linearly with thread count like the current implementation.

Change-Id: I061dcdf5ed93efced8aa0236d0b6a4b9f40232d6