BUG: Fix compilation issues for StatisticsUniqueLabelMapFilterTest
After the corresponding patch was merged with the master, some of the
compilers (namely, MacOSX-Xcode6, MacOSX-XCode6-debug, and
Linux-x86_64-gcc4.5) complained about the
itkStatisticsUniqueLabelMapFilterTest2 test being failing when comparing
the baseline and the test output. It seems like the output had a
difference of some pixels. This patch adds a tolerance of some pixels to
the comparison. Although the issue may be due to a difference in
numerical precision of the mentioned compilers, this may need further
investigation.

Change-Id: I96a321919e707ff22d6c011135de6bdb15443819