BUG: Test case was not verifying meta-data
The test cases used determined that all pixel
values were the same, but did not check that
the image meta-data was preserved during
writing!

The tests have been improved to demonstrate that
a bug has existed in the MGHIO writer for a very
long time.  If the DirectionCosine = Transpose(DirectionCosine)
then there was not problem.

A future patch will address the bug.