COMP: Splitting up IOMINC module into two parts: MINC and Transform
This is to address problem when compiling ITKv4
with older version of gcc (4.4). When MINC IO works, but
IOTransformMINC does not get registered inside ITK hierarchy for some
strange reason. Happens on Debian 6, Ubuntu 10.04 (gcc 4.4), but not on
CentOS 6.
With this change the ITKIOTransformMINC goes into a separate module
named ITKIOTransformMINC  which doesn't have ENABLED_SHARED flag set,
similar to other IOTransform modules.

Change-Id: I676e5927a2b3a0c33a84d81527381b074ad50787