BUG: Fix comparison issues for itkStatisticsUniqueLabelMapFilter
This patch adds a tolerance of some pixels to the comparison of the output
of the itkStatisticsUniqueLabelMapFilterTest2 test (testing the filter
itkStatisticsUniqueLabelMapFilter) and the baseline. Although a previous
branch was supposed to address the issue by changing the structuring
element from a Ball type to a Box, the test still shows some differences
between the otuput and the baseline across some platforms. Adding a
tolerance will prevent these reportedly false positives stand out in the
dashboards.

Change-Id: Ifa42445b9a7f979fd14395ec1a200445d67b2187