BUG: Test for StatisticsUniqueLabelMapFilter stability
This patch set adds to the test the output of the dilation filter in the
itkStatisticsUniqueLabelMapFilter pipeline. A comparison to the
corresponding baseline is also added to the CMakeLists.txt.
This additional output will help in ruling out the dilation operation, as
it seems to be related to the stability issues of the filter.
This branch is related to the JIRA issue 3370.

Change-Id: I9303d8b9c2058aaa2bb4d6ebcc3daa9066a90dc9