COMP: Set target OS X version in environment for tools using compiler
ITK did not build with python wrapping on OS X unless environment variables were
set.

When building on OS X, CMake's default compiler needs flags or environment
variables to determine what version of the OS X SDK to build against.  For
subprojects such as PCRE, this information needs to be passed to the configure
script as environment variables on the configure will fail.  These environment
variables also need to be set for the build and install steps of these
subprojects.

CastXML also needs these environment variables set so that when it examines the
include path it finds the correct directories and generates code that is valid
for the current target platform.

Change-Id: I47104a472e680ebc7169236a65225b9ad8021bf5