ENH: Do not link against specific Python library on OS X
The goal of this change is to allow the importing interpreter
to provide cpython's symbols; that interpreter may be different
from the one which was used to build ITK.

This is necessary to allow a module to be built with a Python installed
to a particular prefix (i.e. system Python) and imported from
a binary-compatible Python in a different prefix
(i.e. Homebrew or Python.org Python).

target_link_libraries() was passing the PYTHON_LIBRARY down to the linker.
For example, when using the python version coming from homebrew:
-I/usr/local/Cellar/python/2.7.10_2/Frameworks/Python.framework/Versions/2.7/include/python2.7

Trying to use itk from OS X's system Python would then result in the following error:
Fatal Python error: PyThreadState_Get: no current thread
Abort trap: 6

A solution is to not pass the PYTHON_LIBRARY, and use the -undefined dynamic_lookup flag.

I tested this on OS X 10.10 with latest clang. ITK can the be imported from all Pythons with
are ABI compatible (2.7.x in our case). Tested with OS X Python 2.7.6 and homebrew's Python 2.7.10.

I only allowed this for Clang; not sure if gcc supports this.
Maybe some older gcc 4.2 versions on OS X.

OS X supports the flag since 10.3, see:
https://developer.apple.com/library/mac/documentation/DeveloperTools/Conceptual/MachOTopics/1-Articles/executing_files.html

Using otool on the .so modules (otool -L _ITKCommonPython.so) shows
that there is no more reference of the python library when using the
 -undefined dynamic_lookup flag.

This idea was originally proposed by Tim Smith, on the homebrew-science repo.

Change-Id: I82ba98d3169a95b10a77a60ed22929eaac839e71