ENH: Enable base classes for templates of landmark initializer
This patch facilitates wrapping by reducing the number of needed
template pemutations that need to be instantiated to only dimension of
the transform.

First, the base itk::Transform class is enabled as a template parameter
by using dynamic_casts to dispatch the correct method for the
transform type, when the template parameter type does not match and
existing InternalInitializeTransform method.

Second, the images parameters are only used for initializing the
bspline extent. So only an itk::ImageBase is needed for this
interface.

PrintSelf method is clean up with usage of macros, and printing
missing ivars.

Change-Id: I54e8d26f5ef4b8663228a3c9ea9bc5e4c0d163be