ENH: Adds ExactlyEquals and NotExactlyEquals functions to itkMath.h
When comparing scalar values, it is sometimes required to determine an
exact comparison. This can be unexpected especially when an exact
comparison between floating type values or values of slightly different
types are requested. This can cause problems, not only can it be ambiguous
as to weather the developer deliberately requested an exact comparison
but compiler warnings can also be generated. Usage of the ExactlyEquals() and
NotExactlyEquals() functions enhance clarity by remove the ambiguity of the
code's intent, while also suppressing compiler warnings.

Change-Id: I969715a63210f6b93a176b9c9aa9dcb3c22055b0