ENH: Dependencies on itkVariableLengthVector.h reduced
A few files were depending on itkVariableLengthVector.h while it's not
necessary.
This implies a few more recompilations than required each time the file is
changed.

I'd like to remove the dependency from itkLinearInterpolateImageFunction.h,
as well. It would mean to define a new trait dedicated to the Zero
Initialization of pixels. The traits' specialization would be included from
itkVariableLengthVector.h in order to avoid to forget to include it, and in
order to avoid to include itkVariableLengthVector.h from
itkLinearInterpolateImageFunction.h.

Change-Id: Id91051e15f28622a7dd5e9d36d46324f1432ebaa