ENH: Improve code coverage for itkParametricSpaceToImageSpaceMeshFilter
The existing test for the itkParametricSpaceToImageSpaceMeshFilter had a
low code coverage because the PointData for the mesh was not being
specified. Thus, itkParametricSpaceToImageSpaceMeshFilter.hxx:94 was true
and the test returned without actually exercising the filter.

This test can be used as example on how to use this filter, however it is
really important to notice that the dimension of the image and the mesh *MUST*
be the same.

Change-Id: I317e91b11cb94d7915d68d53fe60a26739c5ca39