PERF: Performance enhancements on VLV 2/...
This commit introduces rvalue semantics into VariableLengthVector (copy,
assignment, and arithmetic operations).
This implies this evolution is meant to be compiled with a C++11 compiler.

Given a ressampling test case that takes ~120 seconds with the previous version
of VariableLengthVector code. From this commit on, the computation still takes
~120 seconds when compiled in C++98/03, but between 86 and 90 seconds when
compiled with C++11 enabled.

Change-Id: I4fb3a6eb72af50703564581c12bd0d15471c5b84