ENH: COMP: Adds support for complex values to Math::AlmostEquals
Sometimes approximate comparisons between complex values are needed.
A divide by zero check using the AlmostEquals function with a complex
number in itkDivideImageFilter was causing errors when building
ITK Python Wrapping.
To make this possible, an IsComplex trait was additionally added to
itkNumericTraits which returns true for std::complex values, and false
for others. Tests were also added to test the new functionality.

Change-Id: I81ae0525c525417ab57763e809c7ee9c46ac2d98