ENH: Correctly demonstrate float point Eigen failure
Change to floating point computation to correctly match computation
frequently done with Hessian Analysis. This reproduces the same
erroneous eigen values what were observed in
itkDiscreteHessianGaussianImageFunctionTest

The correct and verified values are:
 [-3.26256000e-06,   1.96703376e+01, -2.74458376e+01]
These are produced in the double computation and by numpy. The
incorrect eigen values computed are:
 [-2.86102e-06, 19.6703, -27.4458]

As Eigen vector are ambiguous as to the sign. The test now adopt the
convention of the first component of the Eigen vector being positive
to uniquely define the vector, and allow the results to change by a
sign.

Change-Id: Ia125a036dac316ae45e7f3a52d456bf07e333c5c