BUG: Use CastXML binary built in Release.
Also link with static MSVC runtime.

Built with VS2013 Community Edition with
https://github.com/thewtex/CastXMLSuperbuild

The -fexceptions flag is necessary to prevent castxml from complaining that
exceptions must be enabled to handle "throw" when castxml runs.  Exceptions were
disabled for MSVC in Clang SVN r241952, which is included in LLVM/Clang 3.7rc01
used in this binary.

ITK-3372

Change-Id: I6a09a5d6b73c0bec7e99d08e11722fa288eb9b90