ENH: NEW: Python Wrapping for MinimalPathExtraction
Enables the MinimalPathExtraction module to be wrapped for Python.
-Definition of class PathInfo has been removed from SpeedFunctionToPathFilter.h
and is now known as SpeedFunctionPathInformation with its own header files. This
enables to create an instance of this class in python.
-Wrapping files have been added for each class
-Instanciation of Path Information has been modified in MinimalPathTest.h and
SpeedFunctionToPathFilter header files in order to fit the new definition of
the class.
-ImageToPathFilter header files have been removed because they already exist
in ITK/Module/Nonunit/Review and are already wrapped there. Another commit on
the ITK git repository will be created to move those files in the Path module.